requirements.txt
django==4.0
gunicorn==20.1.0
Pillow
numpy
django-bootstrap4 
mysqlclient
requests